did refactoring for create, delete tests for orders

+++ src/tests/api/orders/delete.spec.ts
@@ -1,7 +1,6 @@
 import { test } from "fixtures";
 import { STATUS_CODES } from "data/statusCodes";
 import { validateResponse } from "utils/validation/validateResponse.utils";
-import { createOrderTestData, IOrderTestData } from "utils/orders/createOrderTestData.utils";
 import { TAGS } from "data/tags";
 import { DELETE_ORDER_CASES } from "data/salesPortal/orders/createOrderTestData";
 import { TIMEOUT_30_S } from "data/salesPortal/constants";
@@ -9,7 +8,6 @@ test.setTimeout(TIMEOUT_30_S);
 
 test.describe("[API][Orders][Delete Order]", () => {
   let token = "";
-  let orderId = "";
   let customerId = "";
   let productIds: string[] = [];
 
@@ -17,43 +15,28 @@ test.describe("[API][Orders][Delete Order]", () => {
     token = await loginApiService.loginAsAdmin();
   });
 
-  test.afterEach(async ({ customersApiService, productsApiService }) => {
-    if (productIds.length) {
-      await Promise.all(productIds.map((id) => productsApiService.delete(token, id)));
-      productIds = [];
-    }
+  test.beforeEach(() => {
+    customerId = "";
+    productIds = [];
+  });
 
-    if (customerId) {
-      await customersApiService.delete(token, customerId);
-      customerId = "";
-    }
+  test.afterEach(async ({ customersApiService, productsApiService }) => {
+    await Promise.allSettled(productIds.map((id) => productsApiService.delete(token, id)));
+    if (customerId) await customersApiService.delete(token, customerId);
   });
 
   for (const positiveCase of DELETE_ORDER_CASES) {
     test(
       positiveCase.title,
       { tag: [TAGS.REGRESSION, TAGS.API, TAGS.ORDERS] },
-      async ({ ordersApi, ordersApiService, customersApiService, productsApiService }) => {
-        const data: IOrderTestData = await createOrderTestData({
-          token,
-          customersApiService,
-          productsApiService,
-          productsCount: positiveCase.productsCount,
-        });
-
-        customerId = data.customerId;
-        productIds = data.productIds;
-
-        const createdOrder = await ordersApiService.create(token, customerId, productIds);
-        orderId = createdOrder._id;
-
-        const deleteOrderResponse = await ordersApi.delete(token, orderId);
-
+      async ({ ordersApi, ordersApiService }) => {
+        const order = await ordersApiService.createOrderAndEntities(token, positiveCase.productsCount);
+        customerId = order.customer._id;
+        productIds = order.products.map((product) => product._id);
+        const deleteOrderResponse = await ordersApi.delete(token, order._id);
         validateResponse(deleteOrderResponse, {
           status: STATUS_CODES.DELETED,
         });
-
-        orderId = "";
       },
     );
   }
